feat: Added bikeshare script file
The bikeshare.py script contains a program which Asks user to specify
a city, month, and day to analyze.
A function which loads data for the specified city and filters
by month and day if applicable.
A function which displays statistics on the most frequent times of travel.
A function which displays statistics on the most popular stations and trip.
A function which displays statistics on the total and average trip duration.
A function to display statistics on bikeshare users.
A function to display raw data of trips